frontend/src/features/pulse-sync/backend/requirements.txt
fastapi
uvicorn
textblob
google-auth
google-api-python-client
pytz
